course1/claims-workflow/pyproject.toml
[project]
name = "claims-workflow"
version = "0.1.0"
description = "Claims processing workflow with LangGraph"
requires-python = ">=3.11"

dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]